Rename e2e test file names
