+++ packages/neuron-wallet/tests-e2e/operations/index.ts
@@ -1,9 +1,9 @@
-import { createWallet } from './createWallet'
-import { importWallet } from './importWallet'
-import { searchWalletIndex } from './searchWalletIndex'
-import { currentWalletAddress } from './currentWalletAddress'
-import { sendTransaction } from './sendTransaction'
-import { checkNetworkStatus } from './checkNetworkStatus'
+import { createWallet } from './create-wallet'
+import { importWallet } from './import-wallet'
+import { searchWalletIndex } from './search-wallet-index'
+import { currentWalletAddress } from './current-wallet-address'
+import { sendTransaction } from './send-transaction'
+import { checkNetworkStatus } from './check-network-status'
 
 export {
   createWallet,
