+++ packages/neuron-wallet/tests-e2e/tests/index.ts
@@ -1,8 +1,8 @@
 import Wallet from './wallet'
 import Network from './network'
-import SendTransaction from './sendTransaction'
+import SendTransaction from './send-transaction'
 import Notification from './notification'
-import AddressBook from './addressBook'
+import AddressBook from './address-book'
 
 export default {
   Wallet,
